+++ index.js
@@ -137,7 +137,7 @@ function lookup(path) {
 
 function populateMaps(extensions, types) {
   // source preference (least -> most)
-  var preference = ['apache', undefined, 'iana']
+  var preference = ['nginx', 'apache', undefined, 'iana']
 
   Object.keys(db).forEach(function forEachMimeType(type) {
     var mime = db[type]