+++ HISTORY.md
@@ -3,6 +3,7 @@ unreleased
 
   * Fix incorrectly treating extension-less file name as extension
     - i.e. `'path/to/json'` will no longer return `application/json`
+  * Improve generation of extension to MIME mapping
   * Refactor internals for readability and no argument reassignment
 
 2.0.14 / 2015-06-06