+++ HISTORY.md
@@ -6,6 +6,7 @@ unreleased
   * Improve generation of extension to MIME mapping
   * Refactor internals for readability and no argument reassignment
   * Prefer `application/*` MIME types from the same source
+  * Prefer any type over `application/octet-stream`
 
 2.0.14 / 2015-06-06
 ===================